+++ src/pages/dashboard/Legal.tsx
@@ -6,8 +6,8 @@ import { IoAdd } from "react-icons/io5";
 import { MdDeleteForever, MdOutlineEdit } from "react-icons/md";
 import { API_URL, deleteCase, getLegalList } from "../../api";
 import { useEffect, useState } from "react";
-import { Select } from "@mantine/core";
-
+import { Input, Select } from "@mantine/core";
+import { CiSearch } from "react-icons/ci";
 const Legal = () => {
   const [legalList, setLegalList] = useState<any[]>([]);
   const [filteredLegalList, setFilteredLegalList] = useState<any[]>([]);
@@ -17,6 +17,7 @@ const Legal = () => {
   const [uniqueOwner, setUniqueOwner] = useState<any[]>([])
   const [uniqueUnit, setUniqueUnit] = useState<any[]>([])
   const [uniqueProperty, setUniqueProperty] = useState<any[]>([])
+  const [searchValue, setSearchvalue] = useState<string | null>(null);
 
   const headers = [
     "Sr. No",
@@ -25,6 +26,7 @@ const Legal = () => {
     "Location",
     "Owner",
     "Reason",
+    "Status",
     "Actions ",
   ];
 
@@ -61,20 +63,32 @@ const Legal = () => {
   }, [legalList])
   const applyFilters = () => {
     const filteredData = legalList.filter((item) => {
+      const matchesSearch = !searchValue ||
+        item?.custom_property.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_unit_no.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_location__area.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_legal_reason.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.owner.toLowerCase().includes(searchValue.toLowerCase())
+
+     
+
       const matchesProperty =
         !selectedProperty || item.custom_property === selectedProperty;
       const matchesUnit =
         !selectedUnit || item.custom_unit_no === selectedUnit;
       const matchesCustomer =
         !selectedOwner || item.owner === selectedOwner;
-      return matchesProperty && matchesUnit && matchesCustomer;
+      return matchesSearch && matchesProperty && matchesUnit && matchesCustomer;
     });
     setFilteredLegalList(filteredData);
   };
 
   useEffect(() => {
     applyFilters();
-  }, [selectedProperty, selectedUnit, selectedOwner]);
+  }, [selectedProperty, selectedUnit, selectedOwner, searchValue]);
+  const handleSearchValue = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchvalue(e.target.value);
+  };
   const selectStyle = {
     input: {
       border: "1px solid gray",
@@ -88,7 +102,12 @@ const Legal = () => {
       color: "#000",
     },
   };
-
+  const searchStyle = {
+    input: {
+      border: "1px solid gray",
+      width: "20vw",
+    },
+  };
   return (
     <main>
       <div className="flex">
@@ -136,6 +155,13 @@ const Legal = () => {
                   onChange={(value) => setSelectedOwner(value)}
                   styles={selectStyle}
                 />
+                <Input
+                  onChange={handleSearchValue}
+                  value={searchValue}
+                  styles={searchStyle}
+                  placeholder="search"
+                  leftSection={<CiSearch size={16} />}
+                />
               </div>
             </div>
             <div className="my-4 p-4">
@@ -170,6 +196,7 @@ const Legal = () => {
                             {item.owner}
                           </td>
                           <td className="p-2 py-3">{item.custom_legal_reason}</td>
+                          <td className="p-2 py-3">{item.custom_status_legal}</td>
 
                           <td className="p-2 py-3">
                             <div className="flex gap-3">
@@ -184,8 +211,11 @@ const Legal = () => {
                                 />
                               </Link>
                               <button className="bg-[#F7F7F7] border border-[#C3C3C3] p-1.5 rounded cursor-pointer" onClick={async () => {
-                                await deleteCase(item.name)
-                                getData()
+                               const confirmed = window.confirm(`Are you sure you want to delete this ${item.custom_property}?`);
+                               if (confirmed) {
+                                 await deleteCase(item.name);
+                                 getData();
+                               }
                               }}>
                                 <MdDeleteForever
                                   size={20}
