+++ src/pages/dashboard/MoveIn.tsx
@@ -6,7 +6,8 @@ import { IoAdd } from "react-icons/io5";
 import { MdDeleteForever, MdOutlineEdit } from "react-icons/md";
 import { API_URL, deleteCase, getMoveInList } from "../../api";
 import { useEffect, useState } from "react";
-import { Select } from "@mantine/core";
+import { Input, Select } from "@mantine/core";
+import { CiSearch } from "react-icons/ci";
 const MoveIn = () => {
   const [moveInList, setMoveInList] = useState<any[]>([]);
   const [filteredMoveInList, setFilteredMoveInList] = useState<any[]>([]);
@@ -16,6 +17,7 @@ const MoveIn = () => {
   const [uniqueCustomer, setUniqueCustomer] = useState<any[]>([])
   const [uniqueUnit, setUniqueUnit] = useState<any[]>([])
   const [uniqueProperty, setUniqueProperty] = useState<any[]>([])
+  const [searchValue, setSearchvalue] = useState<string | null>(null);
 
   const headers = [
     "Sr. No",
@@ -61,20 +63,34 @@ const MoveIn = () => {
 
   const applyFilters = () => {
     const filteredData = moveInList.filter((item) => {
+      const matchesSearch = !searchValue ||
+        item?.custom_property.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_unit_no.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_customer.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_start_date.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_end_date.toLowerCase().includes(searchValue.toLowerCase()) ||
+        item?.custom_statusmi.toLowerCase().includes(searchValue.toLowerCase())
+
+
       const matchesProperty =
         !selectedProperty || item.custom_property === selectedProperty;
       const matchesUnit =
         !selectedUnit || item.custom_unit_no === selectedUnit;
       const matchesCustomer =
         !selectedCustomer || item.custom_customer === selectedCustomer;
-      return matchesProperty && matchesUnit && matchesCustomer;
+      return matchesSearch && matchesProperty && matchesUnit && matchesCustomer;
     });
     setFilteredMoveInList(filteredData);
   };
 
   useEffect(() => {
     applyFilters();
-  }, [selectedProperty, selectedUnit, selectedCustomer]);
+  }, [selectedProperty, selectedUnit, selectedCustomer, searchValue]);
+
+  const handleSearchValue = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchvalue(e.target.value);
+  };
+
   const selectStyle = {
     input: {
       border: "1px solid gray",
@@ -86,6 +102,12 @@ const MoveIn = () => {
     dropdown: {
       backgroundColor: "#F5F5F5",
       color: "#000",
+    },
+  };
+  const searchStyle = {
+    input: {
+      border: "1px solid gray",
+      width: "20vw",
     },
   };
   return (
@@ -135,6 +157,13 @@ const MoveIn = () => {
                   onChange={(value) => setSelectedCustomer(value)}
                   styles={selectStyle}
                 />
+                <Input
+                  onChange={handleSearchValue}
+                  value={searchValue}
+                  styles={searchStyle}
+                  placeholder="search"
+                  leftSection={<CiSearch size={16} />}
+                />
               </div>
             </div>
             <div className="my-4 p-4">
@@ -163,16 +192,9 @@ const MoveIn = () => {
                         <td className="p-2 py-3">{item.custom_start_date}</td>
                         <td className="p-2 py-3">{item.custom_end_date}</td>
                         <td className="p-2 py-3">
-                          <div
-                            className={`p-1 rounded ${item.custom_status === "Draft"
-                              ? "bg-red-400 text-black"
-                              : item.custom_status === "Active"
-                                ? "bg-[#34A853] text-white"
-                                : "bg-blue-400 text-white"
-                              }`}
-                          >
-                            {item.custom_status}
-                          </div>
+
+                          {item.custom_statusmi}
+
                         </td>
                         <td className="p-2 py-3">
                           <div className="flex gap-3">
@@ -187,8 +209,11 @@ const MoveIn = () => {
                               />
                             </Link>
                             <button className="bg-[#F7F7F7] border border-[#C3C3C3] p-1.5 rounded cursor-pointer" onClick={async () => {
-                              await deleteCase(item.name)
-                              getData()
+                             const confirmed = window.confirm(`Are you sure you want to delete this ${item.custom_property}?`);
+                             if (confirmed) {
+                               await deleteCase(item.name);
+                               getData();
+                             }
                             }}>
                               <MdDeleteForever
                                 size={20}
